% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bowersEWS}
\alias{bowersEWS}
\title{Summary of the Accuracy of Early Warning Indicators}
\source{
Bowers, A.J., Sprott, R., Taff, S.A. (2013) Do we Know Who Will Drop Out? A Review of
the Predictors of Dropping out of High School: Precision, Sensitivity and Specificity. The
High School Journal, 96(2), 77-100. \url{doi:10.1353/hsj.2013.0000} (Preprint available.)
}
\description{
This dataset contains accuracy indicators for 110 separate predictive indicators
of student dropout. These indicators are drawn from a literature search of over
30 separate studies in American public schools of early indicators of dropout.
}
\references{
Alex J. Bowers: \url{http://www.tc.columbia.edu/academics/?facid=ab3764}
}
\keyword{data}

