% Generated by roxygen2 (4.0.0): do not edit by hand
\name{buildDISFrame}
\alias{buildDISFrame}
\title{Generate an empty dataframe to match \code{\link{modAcc}} lists}
\usage{
buildDISFrame(methods)
}
\arguments{
  \item{methods}{a list of \code{train} method names to
  generate the dataframe for}
}
\value{
a \code{\link{data.frame}} with the following columns:
\itemize{ \item{DIST - the distance from perfect
classification} \item{DISTSD - the standard deviation of
the distance} \item{SENS - The sensitivity of the
classifier} \item{SPEC - The specificity of the classifier}
\item{method - the area under the curve} \item{grp - the
area under the curve} \item{elapsedTime - the time reported
for the model to run} }
}
\description{
Used for generating the data to make comparative distance
statistics of binary classifiers.
}
\note{
The measures come from the \code{\linkS4class{DISit}}
object
}

