% Generated by roxygen2 (4.0.1): do not edit by hand
\name{buildROCcurveFrame}
\alias{buildROCcurveFrame}
\title{Generate an empty dataframe to match \code{\link{modAcc}} lists}
\usage{
buildROCcurveFrame(methods)
}
\arguments{
  \item{methods}{a list of \code{train} method names to
  generate the dataframe for}
}
\value{
a \code{\link{data.frame}} with the following columns:
\itemize{ \item{sens - the sensitivities of the model at
various thresholds} \item{spec - the specificities of the
model at various thresholds} \item{auc - the area under the
curve} \item{auc - the standard deviation of the AUC for
the best method reported by train} \item{method - the model
method} \item{grp - whether the model is using training or
test data} \item{elapsedTime - the time reported for the
model to run} }
}
\description{
Used for generating the data to make good looking ROC
curves of training and test data.
}
\note{
The sensitivities and specificities come from the
\code{\link{roc}} object stored in the
\code{\linkS4class{ROCit}} object
}

