% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ci}
\alias{ci}
\title{Calculate a confidence interval for a vector}
\usage{
ci(x, scale)

ci(x, scale)
}
\arguments{
  \item{x}{a numeric vector, can contain missing values, if
  so, they are ignored}

  \item{scale}{the factor to multiply the standard
  deviation by when creating the interval}

  \item{x}{A numeric vector which contain NAs}

  \item{scale}{The scale factor of the standard deviation
  out of which to build the CI}
}
\value{
a numeric vector with three elements, the low end of the
confidence interval, the median, and the high end of the
confidence interval
}
\description{
Given a vector, this function produces a three number
summary consisting of the median, the median minus the
standard deviation (multiplied by a scaling factor), and
the median plus the standard deviation multiplied by a
scaling factor.

Given a vector of numerics, x, this returns a numeric
vector of length 3, low confidence interval, high
confidence interval, and the median
}

