% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{cleanArg}
\alias{cleanArg}
\title{An internal function to ensure all parameters passed through \code{\link{modSearch}}
are passed properly}
\usage{
cleanArg(args)
}
\arguments{
  \item{args}{a list extracted from ...}
}
\description{
Reevaluates the ... arguments in the global environment to
extract the proper values
}

