\name{confuse_mat}
\alias{confuse_mat}
\title{An internal function for creating the confusion matrix for \code{\linkS4class{ROCit}}
objects}
\usage{
confuse_mat(mod, thresh, prop = FALSE, testdata)
}
\description{
An internal function for creating the confusion matrix for
\code{\linkS4class{ROCit}} objects
}

