% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dfExtract}
\alias{dfExtract}
\title{Generate a dataframe from \code{\link{modAcc}} lists}
\usage{
dfExtract(mod)
}
\arguments{
  \item{mod}{a list resulting from a call to
  \code{\link{modAcc}}}
}
\value{
a data.frame with columns relating to RMSE or ROC fits
}
\description{
Generic function to extract either ROC data or RMSE data
from models
}
\note{
See dfExtractRMSE or dfExtractROC for details
}

