% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dfExtractRMSE}
\alias{dfExtractRMSE}
\title{Generate a dataframe from \code{\link{modAcc}} lists}
\usage{
dfExtractRMSE(mod)
}
\arguments{
  \item{mod}{a list resulting from a call to
  \code{\link{modAcc}}}
}
\value{
a \code{\link{data.frame}} with the following columns:
\itemize{ \item{RMSE - the RMSE for the model on the test
or training data} \item{RMSEsd - the standard deviation of
the RMSE reported by training with caret} \item{Rsquare -
The R squared reported for the best-fit in caret}
\item{RsquareSD - the model method} \item{method - the area
under the curve} \item{grp - the area under the curve}
\item{elapsedTime - the time reported for the model to run}
}
}
\description{
Used for generating the data to make good looking RMSE
plots for training and test data simultaneously
}
\note{
The measures come from the \code{\linkS4class{RMSEit}}
object
}

