% Generated by roxygen2 (4.0.0): do not edit by hand
\name{DIStest.train}
\alias{DIStest}
\alias{DIStest.train}
\title{Getting an DIStest on a train object}
\usage{
\method{DIStest}{train}(mod, testdata, ...)

DIStest(mod, testdata, ...)
}
\arguments{
  \item{mod}{A model object to generate an
  \code{\linkS4class{DISit}} for}

  \item{testdata}{A dataframe to generate the ROC for the
  mode on}

  \item{...}{optional additional parameters.}
}
\value{
A \code{\linkS4class{DISit}} object
}
\description{
Getting an DIStest on a train object

Explore the accuracy of binary classifiers using distance
from a perfect classifier
}
\details{
The object has the following items \itemize{ \item{thresh -
the threshold for the ROC} \item{auc - the area under the
curve} \item{confusematrix - the confusion matrix for the
ROC fit} \item{rarepercent - percent of rare class correct}
\item{falsepositive - percent of false rare class
identifications} \item{modtype - the class of the model
object} \item{modcall - the call to the model fitting
function} \item{datatype - whether the ROC was computed on
the "train" or the "test" data} }
}
\note{
Yadda yadda yadda
}
\author{
Jared E. Knowles
}

