\name{EWStools-package}
\alias{EWStools-package}
\alias{EWStools}
\docType{package}
\title{
Tools for Building and Testing Early Warning Systems for Education
}
\description{
Tools to help education data analysts build early warning metrics for educational 
outcomes. 
}
\details{
\tabular{ll}{
Package: \tab EWStools\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-04-01\cr
License: \tab GPL-2 \cr
}

This package is built around improving the ability to develop ROC metrics for 
classification

}
\author{
Created by Jared E. Knowles

Maintainer: Jared E. Knowles <jknowles@gmail.com>

}
\references{
Bowers, A.J., Sprott, R., Taff, S.A. (2013) Do we Know Who Will Drop Out? A Review of 
the Predictors of Dropping out of High School: Precision, Sensitivity and Specificity. The 
High School Journal, 96(2), 77-100. \url{doi:10.1353/hsj.2013.0000} (Preprint available.)
}
\keyword{ package }
\seealso{
 \code{\link[caret-package]{caret}} 
}
\examples{
~~ simple examples of the most important functions ~~
}
