% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{factor_norm}
\alias{factor_norm}
\title{Internal function to aide with factors in predicting new data for ROCtest}
\usage{
factor_norm(mod, testdata, impute = FALSE, ...)
}
\description{
Internal function to aide with factors in predicting new data for ROCtest
}
\keyword{internal}

