% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fourStatsSummary}
\alias{fourStatsSummary}
\title{Calculate distance from perfect prediction on resamples}
\usage{
fourStatsSummary(data, lev = levels(data$obs), model = NULL)
}
\arguments{
  \item{data}{a data frame or matrix with columns obs and
  pred for the observed and predicted outcomes. For
  twoClassSummary, columns should also include predicted
  probabilities for each class. See the classProbs argument
  to \code{\link{trainControl}}}

  \item{lev}{a character vector of factors levels for the
  response}

  \item{model}{a character string for the model name}
}
\description{
Calculate distance from perfect prediction on resamples
}
\author{
Max Kuhn
}
\references{
\url{http://appliedpredictivemodeling.com/blog/?offset=1393611546982}
}

