% Generated by roxygen2 (4.0.0): do not edit by hand
\name{modSearchResults}
\alias{modSearchResults}
\title{Summarize the results of modSearch}
\usage{
modSearchResults(df, n)
}
\arguments{
  \item{df}{a dataframe resulting from a call to
  \code{\link{modSearch}}}

  \item{n}{an integer specified by the user representing
  how many "best" and "worst" models to report}
}
\value{
a list with a bunch of elements
}
\description{
Given a dataframe produced by modSearch, identify the top
and bottom n models and describe their accuracy and their
efficiency of fit
}

