% Generated by roxygen2 (4.0.1): do not edit by hand
\name{modTest}
\alias{modTest}
\title{Train a model and store \code{\linkS4class{ROCit}} tests on different datasets}
\usage{
modTest(method, datatype = c("train", "test"), traindata, testdata,
  modelKeep = FALSE, length, fitControl = NA, metric = "ROC",
  cores = NA, ...)
}
\arguments{
  \item{method}{a a string specifying which classification
  or regression model to use. Possible values are found
  using \code{names(getModelInfo())}. See
  http://caret.r-forge.r-project.org/bytag.html. A list of
  funciotns can also be passed for a custom model function.
  See http://caret.r-forge.r-project.org/custom_models.html
  for details}

  \item{datatype}{a named character representing the
  accuracy object be built on either "train" or "test"
  data, user can include both}

  \item{traindata}{a list of length two containing a named
  slot for the matrix of predictors (pred) and the vector
  of classes (class)}

  \item{testdata}{a list of length two containing a named
  slot for the matrix of predictors (pred) and the vector
  of classes (class)}

  \item{modelKeep}{a logical indicating whether the
  original model object should be stored}

  \item{length}{an integer denoting the number of levels of
  each tuning parameter that should be generated to be
  passed to \code{tuneLength} in the \code{train} call}

  \item{fitControl}{an object generated by
  \code{trainControl} to control the behavior of
  \code{train}. If none is given a default is selected.}

  \item{metric}{a character string passed to \code{train}.
  a string that specifies what summary metric will be used
  to select the optimal model. By default, possible values
  are "RMSE" and "Rsquared" for regression and "Accuracy"
  and "Kappa" for classification. If custom performance
  metrics are used (via the \code{summaryFunction} argument
  in trainControl, the value of metric should match one of
  the arguments. If it does not, a warning is issued and
  the first metric given by the summaryFunction is used.}

  \item{cores}{An integer representing the number of cores
  to use on Windows. If not on windows, a warning is
  issued.}

  \item{...}{Additional arguments to be passed to
  \code{\link{train}}}
}
\value{
A character string with an error if unsuccessful. The
result of the \code{modAcc} call if successful: \itemize{
\item{method - the \code{\link{train}} method used to fit
the model} \item{summaryTr - the \code{\linkS4class{ROCit}}
for the training data} \item{summaryTe - the
\code{\linkS4class{ROCit}} for the test data} \item{time -
the time reported for the model to run, taken from the
\code{\link{train}} object if available} }
}
\description{
This function wraps the \code{train} function in the
\code{caret} package with model accuracy reports. It also
allows for errors in fitting models to be caught to make it
easier to use in a loop.
}
\note{
The values presented are for the optimal threshold as
computed by the \code{\link{roc}} function. For some model
types linear combos of predictors may be omitted.
}

