% Generated by roxygen2 (4.0.0): do not edit by hand

\name{print.ROCit}
\alias{print.ROCit}
\title{Print a summary of an \code{\linkS4class{ROCit}} object}
\usage{
\method{print}{ROCit}(x, ...)
}
\arguments{
  \item{x}{a \code{\linkS4class{ROCit}} object generated by
  \code{\link{ROCtest}}}

  \item{...}{optional additional parameters.}
}
\value{
A
}
\description{
Print a summary of an \code{\linkS4class{ROCit}} object
}
\note{
The values presented are for the optimal threshold as
computed by the \code{\link{roc}} function.
}

