% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{probExtract}
\alias{probExtract}
\title{Extract probabilities from various types of R models used for classification}
\usage{
probExtract(mod, testdata = NULL)
}
\arguments{
\item{mod}{an R model object either of glm, caret, or caretEnsemble class}

\item{testdata}{a list with two elements, a preds dataframe and a factor called class}
}
\value{
a dataframe with yhat, the inverse of yhat, and the outcome as a factor
}
\description{
Extracts probabilities consistently for using in confusion matrix calculations
}
\details{
Only works for caretEnsemble, train, and glm objects currently.
}

