% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{reclassProb}
\alias{reclassProb}
\title{Convert probabilities to classifications}
\usage{
reclassProb(yhats, thresh)
}
\arguments{
  \item{yhats}{a dataframe resulting from a call to
  \code{\link{probExtract}}} \item{thresh}{a numeric from 0 to
  1 representing the cutpoint for the probability}
}
\value{
a vector the length of rows in yhats representing the
conversion from a numeric probability to a factor using the
threshold thresh
}
\description{
For building confusion matrices
}
\details{
Only works for caretEnsemble, train, and glm objects
currently. Corrects for the differential ordering in what
the predicted probability is with respect to.
}
\note{
If you find errors, it is likely that the probabilities are
inverted. Please report this.
}
\keyword{internal}

