% Generated by roxygen2 (4.0.0): do not edit by hand
\name{RMSEtest}
\alias{RMSEtest}
\alias{RMSEtest.train}
\title{Generic function to build RMSEtest}
\usage{
RMSEtest(mod, testdata, ...)

\method{RMSEtest}{train}(mod, testdata, ...)
}
\arguments{
  \item{mod}{A model object to generate an
  \code{\linkS4class{RMSEit}} for}

  \item{testdata}{A dataframe to generate the ROC for the
  mode on}

  \item{...}{optional additional parameters.}
}
\value{
A \code{\linkS4class{RMSEit}} object
}
\description{
Allow the user to specify a formula for generating a
binomial dependent variable

Getting an RMSEtest on a train object
}
\details{
The object has the following items \itemize{ \item{bestFit
- best results from train} \item{RMSE - best RMSE}
\item{datatype - whether the ROC was computed on the
"train" or the "test" data} \item{modtype - the class of
the model object} \item{modcall - the call to the model
fitting function} }
}
\note{
Yadda yadda yadda
}
\author{
Jared E. Knowles
}

