\name{ROCtest.glm}
\alias{ROCtest.glm}
\title{Getting an ROCtest on a generalized linear model}
\usage{
\method{ROCtest}{glm}(mod, testdata, ...)
}
\arguments{
  \item{mod}{A model object to generate an
  \code{\linkS4class{ROCit}} for}

  \item{testdata}{A dataframe to generate the ROC for the
  mode on}

  \item{...}{optional additional parameters.}
}
\description{
Getting an ROCtest on a generalized linear model
}

