\name{ROCtest.train}
\alias{ROCtest.train}
\title{Getting an ROCtest on a train object}
\usage{
\method{ROCtest}{train}(mod, testdata, ...)
}
\arguments{
  \item{mod}{A model object to generate an
  \code{\linkS4class{ROCit}} for}

  \item{testdata}{A dataframe to generate the ROC for the
  mode on}

  \item{...}{optional additional parameters.}
}
\description{
Getting an ROCtest on a train object
}

