% Generated by roxygen2 (4.0.0): do not edit by hand
\name{subscores}
\alias{subscores}
\title{Calculate subscores from a dataframe of overall scores}
\usage{
subscores(df, VAR, RISK)
}
\arguments{
  \item{df}{a vector of character representing method names
  in caret}

  \item{VAR}{the number of clusters}

  \item{RISK}{the method used to calculate the
  dissimilarity}
}
\value{
Returns either a dataframe or a matrix with the
dissimilarity values for the methods sampled and the names
of the methods.
}
\description{
Create a new dataframe of subscores for individual
observations across predictor domains
}

