% Generated by roxygen2 (4.0.0): do not edit by hand
\name{summary.ROCit}
\alias{summary.ROCit}
\title{Extract a summary of an \code{\linkS4class{ROCit}} object}
\usage{
\method{summary}{ROCit}(object, ...)
}
\arguments{
\item{object}{a \code{\linkS4class{ROCit}} object generated by \code{\link{ROCtest}}}

\item{...}{optional additional parameters.}
}
\value{
A list with the following items:
\itemize{
\item{datatype - whether the ROC was computed on the "train" or the "test" data}
\item{auc - the area under the curve}
\item{rarepercent - percent of rare class correct}
\item{falsepositive - percent of false rare class identifications}
\item{threshold - the optimal ROC threshold given by \code{\link{ROCtest}}}
\item{confusematrix - the confusion matrix for the ROC fit}
}
}
\description{
Extract a summary of an \code{\linkS4class{ROCit}} object
}
\note{
The values presented are for the optimal threshold as computed by the \code{\link{roc}} function.
}

