% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/builddata.R
\name{buildModelMatrix}
\alias{buildModelMatrix}
\title{Turn a dataframe into a model matrix for caret functions}
\usage{
buildModelMatrix(data, predvars = NULL, keepNA = FALSE)
}
\arguments{
\item{data}{a dataframe that the user would like to convert to a model matrix}

\item{predvars}{a character vector of the names of predictor variables}

\item{keepNA}{behavior with missing values, defaults to FALSE}
}
\value{
A model matrix
}
\description{
Turn a dataframe into a model matrix for caret functions
}
\note{
Built on the \code{\link{model.matrix}} function. Does not produce an
intercept term. Does not drop collinear factor levels.
}

