% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{caretTags}
\alias{caretTags}
\title{Matrix of caret models by tag}
\source{
Max Kuhn. Contributions from Jed Wing, Steve Weston, Andre
Williams, Chris Keefer, Allan Engelhardt, Tony Cooper,
Zachary Mayer and the R Core Team (2014). caret:
Classification and Regression Training. R package version
6.0-24. http://CRAN.R-project.org/package=caret
}
\description{
Data on caret model types.
}
\author{
Jared Knowles
}
\references{
Caret website: \url{http://caret.r-forge.r-project.org/}
}
\keyword{data}

