% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ci}
\alias{ci}
\title{An function for creating a confidence interval from a vector}
\usage{
ci(x, scale)
}
\arguments{
  \item{x}{A numeric vector which contain NAs}

  \item{scale}{The scale factor of the standard deviation
  out of which to build the CI}
}
\description{
Given a vector of numerics, x, this returns a numeric
vector of length 3, low confidence interval, high
confidence interval, and the median
}

