% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confuse_mat}
\alias{confuse_mat}
\title{An internal function for creating the confusion matrix for \code{\linkS4class{ROCit}}
objects}
\usage{
confuse_mat(df, thresh)
}
\description{
An internal function for creating the confusion matrix for
\code{\linkS4class{ROCit}} objects
}
\keyword{internal}

