\name{confuse_mat.train}
\alias{confuse_mat.train}
\title{An internal function for creating the confusion matrix for \code{\linkS4class{ROCit}}
objects from train}
\usage{
confuse_mat.train(x, thresh)
}
\description{
An internal function for creating the confusion matrix for
\code{\linkS4class{ROCit}} objects from train
}
\keyword{internal}

