\name{print.ROCit}
\alias{print}
\alias{print.ROCit}
\title{Print a summary of an \code{\linkS4class{ROCit}} object}
\usage{
\method{print}{ROCit}(x, ...)
}
\arguments{
  \item{object}{a \code{\linkS4class{ROCit}} object
  generated by \code{\link{ROCtest}}}
}
\value{
A
}
\description{
Print a summary of an \code{\linkS4class{ROCit}} object
}
\note{
The values presented are for the optimal threshold as
computed by the \code{\link{roc}} function.
}

