% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/postestimation.R
\name{subscores}
\alias{subscores}
\title{Calculate subscores from an overall score in a dataframe}
\usage{
subscores(df, VAR, RISK)
}
\arguments{
  \item{df}{a dataframe}

  \item{VAR}{a character representing the name of a
  variable in DF}
}
\value{
RISK a character representing the name of a variable in DF
}
\description{
Given a dataframe, specify a variable to be the overall
risk field and loop through all other fields and score them
on whether the observation is closest to the others on the
overall risk
}

